requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.18
prowler==5.1.0
boto3==1.35.94
python-dotenv==1.0.0
